+++ modules/aws-ebs-csi-driver/main.tf
@@ -1,25 +1,31 @@
-module "aws_ebs_csi_driver_pod_identity" {
+locals {
+  name            = "aws-ebs-csi"
+  namespace       = "aws-ebs-csi-driver"
+  service_account = "ebs-csi-controller-sa"
+}
+
+module "pod_identity" {
   source  = "terraform-aws-modules/eks-pod-identity/aws"
   version = "~> 2.0"
 
-  name = "aws-ebs-csi"
+  name = local.name
 
   attach_aws_ebs_csi_policy = true
 
   associations = {
     this = {
       cluster_name    = var.kubernetes_cluster_name
-      namespace       = "aws-ebs-csi-driver"
-      service_account = "ebs-csi-controller-sa"
+      namespace       = local.namespace
+      service_account = local.service_account
     }
   }
 
   tags = var.tags
 }
 
-resource "helm_release" "aws_ebs_csi_driver" {
-  name       = "aws-ebs-csi-driver"
-  namespace  = "aws-ebs-csi-driver"
+resource "helm_release" "this" {
+  name       = local.name
+  namespace  = local.namespace
   repository = "https://kubernetes-sigs.github.io/aws-ebs-csi-driver"
   chart      = "aws-ebs-csi-driver"
   version    = "2.49.0"
@@ -27,9 +33,9 @@ resource "helm_release" "aws_ebs_csi_driver" {
   create_namespace = true
 
   values = [
-    templatefile("${path.module}/values.yaml", {}),
-    var.custom_values_templatefile != "" ? templatefile(var.custom_values_templatefile, var.custom_values_variables) : ""
+    file("${path.module}/values.yaml"),
+    var.values_overrides
   ]
 
-  depends_on = [module.aws_ebs_csi_driver_pod_identity]
+  depends_on = [module.pod_identity]
 }