+++ modules/descheduler/main.tf
@@ -1,6 +1,11 @@
-resource "helm_release" "descheduler" {
-  name       = "descheduler"
-  namespace  = "descheduler"
+locals {
+  name      = "descheduler"
+  namespace = "descheduler"
+}
+
+resource "helm_release" "this" {
+  name       = local.name
+  namespace  = local.namespace
   repository = "https://kubernetes-sigs.github.io/descheduler"
   chart      = "descheduler"
   version    = "0.33.0"
@@ -8,13 +13,7 @@ resource "helm_release" "descheduler" {
   create_namespace = true
 
   values = [
-    var.custom_values_templatefile != "" ? templatefile(var.custom_values_templatefile, var.custom_values_variables) : ""
-  ]
-
-  set = [
-    {
-      name  = "deschedulerPolicy.evictLocalStoragePods"
-      value = "true"
-    }
+    file("${path.module}/values.yaml"),
+    var.values_overrides
   ]
 }