+++ modules/external-dns/main.tf
@@ -1,8 +1,14 @@
-module "external_dns_pod_identity" {
+locals {
+  name            = "external-dns"
+  namespace       = "external-dns"
+  service_account = "external-dns"
+}
+
+module "pod_identity" {
   source  = "terraform-aws-modules/eks-pod-identity/aws"
   version = "~> 2.0"
 
-  name = "external-dns"
+  name = local.name
 
   attach_external_dns_policy    = true
   external_dns_hosted_zone_arns = [var.route53_zone_arn]
@@ -10,17 +16,17 @@ module "external_dns_pod_identity" {
   associations = {
     this = {
       cluster_name    = var.kubernetes_cluster_name
-      namespace       = "external-dns"
-      service_account = "external-dns"
+      namespace       = local.namespace
+      service_account = local.service_account
     }
   }
 
   tags = var.tags
 }
 
-resource "helm_release" "external_dns" {
-  name       = "external-dns"
-  namespace  = "external-dns"
+resource "helm_release" "this" {
+  name       = local.name
+  namespace  = local.namespace
   repository = "https://kubernetes-sigs.github.io/external-dns"
   chart      = "external-dns"
   version    = "1.19.0"
@@ -28,12 +34,12 @@ resource "helm_release" "external_dns" {
   create_namespace = true
 
   values = [
-    templatefile("${path.module}/values.tftpl", {
+    templatefile("${path.module}/values.yaml", {
       domain      = var.route53_zone_name,
       clusterName = var.kubernetes_cluster_name
     }),
-    var.custom_values_templatefile != "" ? templatefile(var.custom_values_templatefile, var.custom_values_variables) : ""
+    var.values_overrides
   ]
 
-  depends_on = [module.external_dns_pod_identity]
+  depends_on = [module.pod_identity]
 }