+++ modules/external-dns/variables.tf
@@ -1,8 +1,3 @@
-variable "tags" {
-  description = "A map of tags to add to all resources"
-  type        = map(string)
-}
-
 variable "kubernetes_cluster_name" {
   description = "Name of the EKS cluster"
   type        = string
@@ -18,14 +13,13 @@ variable "route53_zone_name" {
   type        = string
 }
 
-variable "custom_values_templatefile" {
-  description = "Custom values templatefile to pass to the helm chart"
-  type        = string
-  default     = ""
+variable "tags" {
+  description = "A map of tags to add to all resources"
+  type        = map(string)
 }
 
-variable "custom_values_variables" {
-  description = "Variables for the custom values templatefile"
-  type        = any
-  default     = {}
+variable "values_overrides" {
+  description = "Values in raw yaml format to pass to helm."
+  type        = string
+  default     = null
 }