+++ modules/metrics-server/main.tf
@@ -1,13 +1,16 @@
-resource "helm_release" "metrics_server" {
-  name       = "metrics-server"
-  namespace  = "metrics-server"
+locals {
+  name      = "metrics-server"
+  namespace = "metrics-server"
+}
+
+resource "helm_release" "this" {
+  name       = local.name
+  namespace  = local.namespace
   repository = "https://kubernetes-sigs.github.io/metrics-server"
   chart      = "metrics-server"
   version    = "3.13.0"
 
   create_namespace = true
 
-  values = [
-    var.custom_values_templatefile != "" ? templatefile(var.custom_values_templatefile, var.custom_values_variables) : ""
-  ]
+  values = [var.values_overrides]
 }