+++ modules/nvidia-gpu-operator/main.tf
@@ -1,6 +1,11 @@
-resource "helm_release" "nvidia_device_plugin" {
-  name       = "gpu-operator"
-  namespace  = "gpu-operator"
+locals {
+  name      = "gpu-operator"
+  namespace = "gpu-operator"
+}
+
+resource "helm_release" "this" {
+  name       = local.name
+  namespace  = local.namespace
   repository = "https://helm.ngc.nvidia.com/nvidia"
   chart      = "gpu-operator"
   version    = "v25.3.0"
@@ -8,7 +13,7 @@ resource "helm_release" "nvidia_device_plugin" {
   create_namespace = true
 
   values = [
-    templatefile("${path.module}/values.yaml", {}),
-    var.custom_values_templatefile != "" ? templatefile(var.custom_values_templatefile, var.custom_values_variables) : ""
+    file("${path.module}/values.yaml"),
+    var.values_overrides
   ]
 }